move required args unit tests to unit test directory
