+++ integration/test_alerts.py
@@ -47,12 +47,3 @@ ALERT_ADVANCED_QUERY_COMMAND = "code42 alerts search --advanced-query '{}'".form
 def test_alert_command_returns_success_return_code(command):
     return_code, response = run_command(command)
     assert return_code == 0
-
-
-@pytest.mark.parametrize(
-    "command", ["{} --advanced-query '{}'".format(ALERT_COMMAND, ADVANCED_QUERY)]
-)
-def test_begin_cant_be_used_with_advanced_query(command):
-    return_code, response = run_command(command)
-    assert return_code == 2
-    assert "--begin can't be used with: --advanced-query" in response[0]
