+++ integration/test_legal_hold.py
@@ -23,41 +23,3 @@ LEGAL_HOLD_COMMAND = "code42 legal-hold"
 def test_alert_rules_command_returns_success_return_code(command):
     return_code, response = run_command(command)
     assert return_code == 0
-
-
-@pytest.mark.parametrize(
-    "command, error_msg",
-    [
-        (
-            "{} add-user --matter-id test-matter-id".format(LEGAL_HOLD_COMMAND),
-            "Missing option '-u' / '--username'.",
-        ),
-        (
-            "{} remove-user --matter-id test-matter-id".format(LEGAL_HOLD_COMMAND),
-            "Missing option '-u' / '--username'.",
-        ),
-        (
-            "{} add-user".format(LEGAL_HOLD_COMMAND),
-            "Missing option '-m' / '--matter-id'.",
-        ),
-        (
-            "{} remove-user".format(LEGAL_HOLD_COMMAND),
-            "Missing option '-m' / '--matter-id'.",
-        ),
-        ("{} show".format(LEGAL_HOLD_COMMAND), "Missing argument 'MATTER_ID'."),
-        (
-            "{} bulk add".format(LEGAL_HOLD_COMMAND),
-            "Error: Missing argument 'CSV_FILE'.",
-        ),
-        (
-            "{} bulk remove".format(LEGAL_HOLD_COMMAND),
-            "Error: Missing argument 'CSV_FILE'.",
-        ),
-    ],
-)
-def test_alert_rules_command_returns_error_exit_status_when_missing_required_parameters(
-    command, error_msg
-):
-    return_code, response = run_command(command)
-    assert return_code == 2
-    assert error_msg in "".join(response)
